+++ src/main/java/pl/betoncraft/betonquest/objectives/LoginObjective.java
@@ -1,0 +1,62 @@
+package pl.betoncraft.betonquest.objectives;
+
+import org.bukkit.Bukkit;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.player.PlayerQuitEvent;
+import pl.betoncraft.betonquest.BetonQuest;
+import pl.betoncraft.betonquest.Instruction;
+import pl.betoncraft.betonquest.api.Objective;
+import pl.betoncraft.betonquest.exceptions.InstructionParseException;
+import pl.betoncraft.betonquest.utils.LogUtils;
+import pl.betoncraft.betonquest.utils.PlayerConverter;
+
+import java.util.logging.Level;
+
+/**
+ * Requires the player to join the server.
+ */
+public class LoginObjective extends Objective implements Listener {
+
+    public LoginObjective(final Instruction instruction) throws InstructionParseException {
+        super(instruction);
+        template = ObjectiveData.class;
+        Bukkit.getPluginManager().registerEvents(this, BetonQuest.getInstance());
+    }
+
+    @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
+    public void onJoin(final PlayerJoinEvent event) {
+        final String playerID = PlayerConverter.getID(event.getPlayer());
+        if (containsPlayer(playerID) && checkConditions(playerID)) {
+            completeObjective(playerID);
+        }
+    }
+
+    @Override
+    public void start() {
+    }
+
+    @Override
+    public void close() {
+        HandlerList.unregisterAll(this);
+        super.close();
+    }
+
+    @Override
+    public void stop() {
+    }
+
+    @Override
+    public String getDefaultDataInstruction() {
+        return "";
+    }
+
+    @Override
+    public String getProperty(final String name, final String playerID) {
+        return "";
+    }
+
+}